+++ async_upnp_client/event_handler.py
@@ -95,8 +95,8 @@ class UpnpEventHandler:
 
         # decode event and send updates to service
         changes = {}
-        body = body.strip().strip("\u0000")
-        el_root = DET.fromstring(body)
+        stripped_body = body.rstrip(" \t\r\n\0")
+        el_root = DET.fromstring(stripped_body)
         for el_property in el_root.findall("./event:property", NS):
             for el_state_var in el_property:
                 name = el_state_var.tag